+++ waspc/data/Generator/templates/sdk/wasp/client/index.ts
@@ -4,7 +4,10 @@ export enum HttpMethod {
 	Get = 'GET',
 	Post = 'POST',
 	Put = 'PUT',
+	Patch = 'PATCH',
 	Delete = 'DELETE',
+	Head = 'HEAD',
+	Options = 'OPTIONS',
 }
 
 // PUBLIC API