+++ waspc/data/packages/spec/src/appSpec.ts
@@ -134,7 +134,15 @@ export type ExecutorOptions = {
   pgBoss: Optional<object>;
 };
 
-export type HttpMethod = "ALL" | "GET" | "POST" | "PUT" | "DELETE";
+export type HttpMethod =
+  | "ALL"
+  | "GET"
+  | "POST"
+  | "PUT"
+  | "PATCH"
+  | "DELETE"
+  | "HEAD"
+  | "OPTIONS";
 
 export type HttpRoute = [HttpMethod, string];