+++ waspc/e2e-tests/test-outputs/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/sdk/wasp/client/vite/plugins/virtualUserModules.ts
@@ -9,7 +9,7 @@ import type { Plugin } from "vite";
  * clientVirtualUserModuleMap["virtual:wasp/user/env"] // => "./src/env"
  */
 const clientVirtualUserModuleMap: { [virtualUserModule: string]: string } = {
-  'virtual:wasp/user/env': './src/env',
+  "virtual:wasp/user/env": "./src/env",
 };
 
 /**