+++ waspc/e2e-tests/test-outputs/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/server/src/actions/customSignup.ts
@@ -1,6 +1,6 @@
 import { prisma } from 'wasp/server'
 
-import { customSignup } from '../../../../../src/features/auth/customSignup'
+import { customSignup } from "../../../../../src/features/auth/customSignup"
 
 
 export default async function (args, context) {