+++ waspc/e2e-tests/test-outputs/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/server/src/actions/voidToStringNoAuth.ts
@@ -1,6 +1,6 @@
 import { prisma } from 'wasp/server'
 
-import { voidToStringNoAuth } from '../../../../../src/rpcTests/operations/definitions'
+import { voidToStringNoAuth } from "../../../../../src/rpcTests/operations/definitions"
 
 
 export default async function (args, context) {