+++ waspc/e2e-tests/test-outputs/snapshots/kitchen-sink-golden/wasp-app/.wasp/out/server/src/routes/operations/customSignup.js
@@ -1,4 +1,4 @@
 import { createAction } from '../../middleware/operations.js'
-import customSignup from '../../actions/customSignup.js'
+import customSignup from "../../actions/customSignup.js"
 
 export default createAction(customSignup)