+++ waspc/e2e-tests/test-outputs/snapshots/kitchen-sink-golden/wasp-app/src/features/auth/auth.wasp.ts
@@ -68,7 +68,7 @@ export const authConfig: Auth = {
     email: {
       userSignupFields: emailUserSignupFields,
       fromField: {
-        name: "Wasp Kitchen Sink",
+        name: 'Wasp\'s "Kitchen Sink"',
         email: "kitchen-sink@wasp.sh",
       },
       emailVerification: {