+++ waspc/e2e-tests/test-outputs/snapshots/wasp-compile-golden/wasp-app/.wasp/out/sdk/wasp/client/vite/virtual-files/files/routes.tsx
@@ -9,7 +9,7 @@ import { Outlet } from "react-router"
 const routesMapping = {
   RootRoute: {
     lazy: async () => {
-      const Component = await import('./src/MainPage').then(m => m.MainPage);
+      const Component = await import("./src/MainPage").then(m => m.MainPage);
 
       return {
         Component: